requirements.txt
langchain
chromadb
python-dotenv

langchain-google-genai
google-generativeai

langchain-community
unstructured[all-docs]
pymupdf

openpyxl
python-docx
python-pptx

pillow
transformers
torch
sentence-transformers

duckduckgo-search

chainlit
